+++ laserchicken/compute_neighbors.py
@@ -30,14 +30,14 @@ def compute_cylinder_neighborhood(environment_pc, target_pc, radius):
              the returned indices also contains the index of the target point.
     """
     avg_points_cyl = (radius * radius * math.pi) * POINT_CLOUD_DIST
-    x = target_pc[point]['x']['data']
+    target_points = target_pc[point]['x']['data']
 
     if len(environment_pc[point]['x']['data']) == 0:
-        for _ in x:
-            yield [[] for _ in x]
+        for _ in target_points:
+            yield []
         return
 
-    cyl_size = avg_points_cyl * np.size(x) * sys.getsizeof(int)
+    cyl_size = avg_points_cyl * np.size(target_points) * sys.getsizeof(int)
     mem_size = virtual_memory().total
 
     print("Cylinder size in Bytes: %s" % cyl_size)
@@ -51,12 +51,12 @@ def compute_cylinder_neighborhood(environment_pc, target_pc, radius):
 
         env_tree = kd_tree.get_kdtree_for_pc(environment_pc)
 
-        for i in range(0, np.size(x), num_points):
-            box_points = np.column_stack(
-                (x[i:min(i + num_points, np.size(x))], y[i:min(i + num_points, np.size(x))]))
-            target_box_tree = cKDTree(
-                box_points, compact_nodes=False, balanced_tree=False)
-            yield target_box_tree.query_ball_tree(env_tree, radius)
+        for range_start in range(0, np.size(target_points), num_points):
+            range_end = min(range_start + num_points, np.size(target_points))
+            box_points = np.column_stack((target_points[range_start:range_end], y[range_start:range_end]))
+            target_box_tree = cKDTree(box_points, compact_nodes=False, balanced_tree=False)
+            for neighborhood in target_box_tree.query_ball_tree(env_tree, radius):
+                yield neighborhood
 
     else:
         print("Start tree creation")
@@ -64,7 +64,8 @@ def compute_cylinder_neighborhood(environment_pc, target_pc, radius):
         print("Done with env tree creation")
         target_tree = kd_tree.get_kdtree_for_pc(target_pc)
         print("Done with target tree creation")
-        yield target_tree.query_ball_tree(env_tree, radius)
+        for neighborhood in target_tree.query_ball_tree(env_tree, radius):
+            yield neighborhood
 
 
 def compute_sphere_neighborhood(environment_pc, target_pc, radius):
@@ -77,24 +78,17 @@ def compute_sphere_neighborhood(environment_pc, target_pc, radius):
     :param radius: search radius for neighbors
     :return: indices of neighboring points from the environment point cloud for each target point
     """
-    neighborhoods = compute_cylinder_neighborhood(
-        environment_pc, target_pc, radius)
-
-    counter = 0  # Target and neighborhood indices are going to be out of sync in loop below.
-    for neighborhood_indices in neighborhoods:
-        result = []
-        for i, _ in enumerate(neighborhood_indices):
-            target_x, target_y, target_z = utils.get_point(target_pc, counter)
-            counter += 1
-            result_indices = []
-            for j in neighborhood_indices[i]:
-                env_x, env_y, env_z = utils.get_point(environment_pc, j)
-                if abs(target_z - env_z) > radius:
-                    continue
-                if (env_x - target_x) ** 2 + (env_y - target_y) ** 2 + (env_z - target_z) ** 2 <= radius ** 2:
-                    result_indices.append(j)
-            result.append(result_indices)
-        yield result
+    neighborhoods = compute_cylinder_neighborhood(environment_pc, target_pc, radius)
+    for target_i, neighborhood_indices in enumerate(neighborhoods):
+        target_x, target_y, target_z = utils.get_point(target_pc, target_i)
+        result_indices = []
+        for neighbor_j in neighborhood_indices:
+            env_x, env_y, env_z = utils.get_point(environment_pc, neighbor_j)
+            if abs(target_z - env_z) > radius:
+                continue
+            if (env_x - target_x) ** 2 + (env_y - target_y) ** 2 + (env_z - target_z) ** 2 <= radius ** 2:
+                result_indices.append(neighbor_j)
+        yield result_indices
 
 
 def compute_cell_neighborhood(environment_pc, target_pc, side_length):
@@ -110,25 +104,19 @@ def compute_cell_neighborhood(environment_pc, target_pc, side_length):
 
     max_radius = 0.5 * math.sqrt((side_length ** 2) + (side_length ** 2))
 
-    neighbors = compute_cylinder_neighborhood(
+    neighborhoods = compute_cylinder_neighborhood(
         environment_pc, target_pc, max_radius)
 
-    counter = 0  # Target and neighborhood indices are going to be out of sync in loop below.
-    for neighborhood_indices in neighbors:
-        result = []
-        for i, _ in enumerate(neighborhood_indices):
-            target_x, target_y, _ = utils.get_point(target_pc, counter)
-            counter += 1
-            neighbor_indices = neighborhood_indices[i]
-            result_indices = []
-            for j in neighbor_indices:
-                env_x, env_y, _ = utils.get_point(environment_pc, j)
-                if ((abs(target_x - env_x)) > 0.5 * side_length) or ((abs(target_y - env_y)) > 0.5 * side_length):
-                    continue
-                else:
-                    result_indices.append(j)
-            result.append(result_indices)
-        yield result
+    for target_i, neighborhood_indices in enumerate(neighborhoods):
+        target_x, target_y, _ = utils.get_point(target_pc, target_i)
+        result_indices = []
+        for neighbor_j in neighborhood_indices:
+            env_x, env_y, _ = utils.get_point(environment_pc, neighbor_j)
+            if ((abs(target_x - env_x)) > 0.5 * side_length) or ((abs(target_y - env_y)) > 0.5 * side_length):
+                continue
+            else:
+                result_indices.append(neighbor_j)
+        yield result_indices
 
 
 def compute_cube_neighborhood(environment_pc, target_pc, side_length):
@@ -142,25 +130,20 @@ def compute_cube_neighborhood(environment_pc, target_pc, side_length):
     :return: indices of neighboring points from the environment point cloud for each target point
     """
 
-    neighbors = compute_cell_neighborhood(
+    neighborhoods = compute_cell_neighborhood(
         environment_pc, target_pc, side_length)
 
-    counter = 0
-    for neighborhood_indices in neighbors:
-        result = []
-        for i, _ in enumerate(neighborhood_indices):
-            _, _, target_z = utils.get_point(target_pc, counter)
-            counter += 1
-            neighbor_indices = neighborhood_indices[i]
-            result_indices = []
-            for j in neighbor_indices:
-                _, _, env_z = utils.get_point(environment_pc, j)
-                if abs(target_z - env_z) > side_length:
-                    continue
-                else:
-                    result_indices.append(j)
-            result.append(result_indices)
-        yield result
+    for target_i, neighborhood_indices in enumerate(neighborhoods):
+        _, _, target_z = utils.get_point(target_pc, target_i)
+        result_indices = []
+        for neighbor_j in neighborhood_indices:
+            _, _, env_z = utils.get_point(environment_pc, neighbor_j)
+            if abs(target_z - env_z) > side_length:
+                continue
+            else:
+                result_indices.append(neighbor_j)
+        yield result_indices
+
 
 
 def compute_neighborhoods(env_pc, target_pc, volume_description, sample_size=None):
@@ -176,27 +159,27 @@ def compute_neighborhoods(env_pc, target_pc, volume_description, sample_size=Non
     volume_type = volume_description.get_type()
 
     if volume_type == Cell.TYPE:
-        neighbors = compute_cell_neighborhood(env_pc, target_pc, volume_description.side_length)
+        neighborhoods = compute_cell_neighborhood(env_pc, target_pc, volume_description.side_length)
     elif volume_type == Cube.TYPE:
-        neighbors = compute_cube_neighborhood(env_pc, target_pc, volume_description.side_length)
+        neighborhoods = compute_cube_neighborhood(env_pc, target_pc, volume_description.side_length)
     elif volume_type == Sphere.TYPE:
-        neighbors = compute_sphere_neighborhood(env_pc, target_pc, volume_description.radius)
+        neighborhoods = compute_sphere_neighborhood(env_pc, target_pc, volume_description.radius)
     elif volume_type == InfiniteCylinder.TYPE:
-        neighbors = compute_cylinder_neighborhood(env_pc, target_pc, volume_description.radius)
+        neighborhoods = compute_cylinder_neighborhood(env_pc, target_pc, volume_description.radius)
     else:
         raise ValueError(
             'Neighborhood computation error because volume type "{}" is unknown.'.format(volume_type))
 
-    for x in neighbors:
-        yield _subsample_if_necessary(x, sample_size)
+    return _subsample_if_necessary(neighborhoods, sample_size)
 
 
-def _subsample_if_necessary(x, sample_size):
+def _subsample_if_necessary(neighborhoods, sample_size):
     if sample_size:
-        sampled = [_subsample(elements, sample_size) if len(elements) > sample_size else elements for elements in x]
+        sampled = (_subsample(neighborhood, sample_size) if len(neighborhood) > sample_size else neighborhood for
+                   neighborhood in neighborhoods)
         return sampled
     else:
-        return x
+        return neighborhoods
 
 
 def _subsample(neighborhood, sample_size=None):