+++ laserchicken/feature_extractor/test_all_features.py
@@ -4,13 +4,12 @@ import os
 import numpy as np
 import pytest
 
-from laserchicken import compute_neighbors, keys, load
+from laserchicken import compute_features, compute_neighborhoods, keys, load
+from laserchicken.keys import point
 #from laserchicken.feature_extractor import *
 from laserchicken.feature_extractor.pulse_penetration_feature_extractor import GROUND_TAGS
-from laserchicken.keys import point, normalized_height
 from laserchicken.utils import copy_point_cloud
-from laserchicken.volume_specification import InfiniteCylinder, Cell
-from .feature_extraction import compute_features
+from laserchicken.volume_specification import InfiniteCylinder
 from .feature_map import create_default_feature_map, _create_name_extractor_pairs
 
 np.random.seed(1234)
@@ -22,15 +21,12 @@ _TEST_DATA_SOURCE = 'testdata'
 _CYLINDER = InfiniteCylinder(4)
 _PC_260807 = load(os.path.join(_TEST_DATA_SOURCE, _TEST_FILE_NAME))
 _PC_1000 = copy_point_cloud(_PC_260807, array_mask=(
-    np.random.choice(range(len(_PC_260807[keys.point]['x']['data'])), size=1000, replace=False)))
+    np.random.choice(range(len(_PC_260807[point]['x']['data'])), size=1000, replace=False)))
 _PC_10 = copy_point_cloud(_PC_260807, array_mask=(
-    np.random.choice(range(len(_PC_260807[keys.point]['x']['data'])), size=10, replace=False)))
-_1000_NEIGHBORHOODS_IN_260807 = next(
-    compute_neighbors.compute_neighborhoods(_PC_260807, _PC_1000, _CYLINDER, sample_size=500))
-_10_NEIGHBORHOODS_IN_260807 = next(
-    compute_neighbors.compute_neighborhoods(_PC_260807, _PC_10, _CYLINDER, sample_size=500))
-_260807_NEIGHBORHOODS_IN_10 = next(
-    compute_neighbors.compute_neighborhoods(_PC_10, _PC_260807, _CYLINDER, sample_size=500))
+    np.random.choice(range(len(_PC_260807[point]['x']['data'])), size=10, replace=False)))
+_1000_NEIGHBORHOODS_IN_260807 = list(compute_neighborhoods(_PC_260807, _PC_1000, _CYLINDER, sample_size=500))
+_10_NEIGHBORHOODS_IN_260807 = list(compute_neighborhoods(_PC_260807, _PC_10, _CYLINDER, sample_size=500))
+_260807_NEIGHBORHOODS_IN_10 = list(compute_neighborhoods(_PC_10, _PC_260807, _CYLINDER, sample_size=500))
 
 features_by_name = create_default_feature_map()
 feature_names = [name for name in features_by_name]
@@ -47,24 +43,16 @@ def test_no_duplicate_feature_registrations():
 @pytest.mark.parametrize("feature", feature_names)
 def test_completeTile_consistentOutput(feature):
     target_point_cloud = copy_point_cloud(_PC_1000)
-    compute_features(copy_point_cloud(_PC_260807), _1000_NEIGHBORHOODS_IN_260807, 0, target_point_cloud,
-                     [feature], volume=_CYLINDER)
+    compute_features(copy_point_cloud(_PC_260807), _1000_NEIGHBORHOODS_IN_260807, target_point_cloud, [feature],
+                     volume=_CYLINDER)
     _assert_consistent_attribute_length(target_point_cloud)
 
 
 @pytest.mark.parametrize("feature", feature_names)
 def test_manyTargets_consistentOutput(feature):
     target_point_cloud = copy_point_cloud(_PC_260807)
-    compute_features(copy_point_cloud(_PC_10), _260807_NEIGHBORHOODS_IN_10, 0, target_point_cloud,
-                     [feature], volume=_CYLINDER)
-    _assert_consistent_attribute_length(target_point_cloud)
-
-
-@pytest.mark.parametrize("feature", feature_names)
-def test_manyTargetsBigEnvironment_consistentOutput(feature):
-    target_point_cloud = copy_point_cloud(_PC_260807)
-    compute_features(copy_point_cloud(_PC_1000), _260807_NEIGHBORHOODS_IN_10, 0, target_point_cloud,
-                     [feature], volume=_CYLINDER)
+    compute_features(copy_point_cloud(_PC_10), _260807_NEIGHBORHOODS_IN_10, target_point_cloud, [feature],
+                     volume=_CYLINDER)
     _assert_consistent_attribute_length(target_point_cloud)
 
 
@@ -72,7 +60,7 @@ def test_manyTargetsBigEnvironment_consistentOutput(feature):
 def test_xAllZeros_consistentOutput(feature):
     n = 10
     pc = _create_point_cloud(x=0, n=n)
-    compute_features(pc, [[] for _ in range(n)], 0, pc, [feature], volume=_CYLINDER)
+    compute_features(pc, [[] for _ in range(n)], pc, [feature], volume=_CYLINDER)
     _assert_consistent_attribute_length(pc)
 
 
@@ -80,7 +68,7 @@ def test_xAllZeros_consistentOutput(feature):
 def test_yAllZeros_consistentOutput(feature):
     n = 10
     pc = _create_point_cloud(y=0, n=n)
-    compute_features(pc, [[] for _ in range(n)], 0, pc, [feature], volume=_CYLINDER)
+    compute_features(pc, [[] for _ in range(n)], pc, [feature], volume=_CYLINDER)
     _assert_consistent_attribute_length(pc)
 
 
@@ -88,7 +76,7 @@ def test_yAllZeros_consistentOutput(feature):
 def test_zAllZeros_consistentOutput(feature):
     n = 10
     pc = _create_point_cloud(z=0, n=n)
-    compute_features(pc, [[] for _ in range(n)], 0, pc, [feature], volume=_CYLINDER)
+    compute_features(pc, [[] for _ in range(n)], pc, [feature], volume=_CYLINDER)
     _assert_consistent_attribute_length(pc)
 
 
@@ -96,7 +84,7 @@ def test_zAllZeros_consistentOutput(feature):
 def test_zeroPoints_consistentOutput(feature):
     n = 0
     pc = _create_point_cloud(n=n)
-    compute_features(pc, [[] for _ in range(n)], 0, pc, [feature], volume=_CYLINDER)
+    compute_features(pc, [[] for _ in range(n)], pc, [feature], volume=_CYLINDER)
     _assert_consistent_attribute_length(pc)
 
 
@@ -118,7 +106,7 @@ def test_twoNeighbors_consistentOutput(feature):
 def _assert_consistent_output_with_n_neighbors(feature, n_neighbors):
     n_points = 10
     pc = _create_point_cloud(n=n_points)
-    compute_features(pc, [range(n_neighbors) for _ in range(n_points)], 0, pc, [feature], volume=_CYLINDER)
+    compute_features(pc, [range(n_neighbors) for _ in range(n_points)], pc, [feature], volume=_CYLINDER)
     _assert_consistent_attribute_length(pc)
 
 
@@ -131,8 +119,7 @@ def test_inputNotChanged(feature):
     original_neighborhoods = _10_NEIGHBORHOODS_IN_260807
     neighborhoods = [[e for e in l] for l in original_neighborhoods]
 
-    compute_features(environment, neighborhoods, 0, targets, [feature],
-                     volume=_CYLINDER)
+    compute_features(environment, neighborhoods, targets, [feature], volume=_CYLINDER)
 
     _assert_attributes_not_changed(original_environment, environment)
     _assert_attributes_not_changed(original_targets, targets)
@@ -157,12 +144,12 @@ def _create_attribute(n_points, fill_value=None):
 
 
 def _assert_attributes_not_changed(original_point_cloud, new_point_cloud):
-    for attribute in original_point_cloud[keys.point]:
-        np.testing.assert_array_almost_equal(new_point_cloud[keys.point][attribute]['data'],
-                                             original_point_cloud[keys.point][attribute]['data'])
+    for attribute in original_point_cloud[point]:
+        np.testing.assert_array_almost_equal(new_point_cloud[point][attribute]['data'],
+                                             original_point_cloud[point][attribute]['data'])
 
 
 def _assert_consistent_attribute_length(target_point_cloud):
-    n_elements = len(target_point_cloud[keys.point]['x'])
-    for key in target_point_cloud[keys.point]:
-        assert n_elements == len(target_point_cloud[keys.point][key])
+    n_elements = len(target_point_cloud[point]['x']['data'])
+    for key in target_point_cloud[point]:
+        assert n_elements == len(target_point_cloud[point][key]['data'])