+++ laserchicken/feature_extractor/test_var_feature_extractor.py
@@ -13,13 +13,13 @@ class TestVariationZFeatureExtractor(unittest.TestCase):
         neighborhood = [89664, 23893, 30638, 128795, 62052, 174453, 29129, 17127, 128215, 29667, 116156, 119157, 98591,
                         7018,
                         61494, 65194, 117931, 62971, 10474, 90322]
-        var_z = self.extractor.extract(pc_in, neighborhood, None, None, None)
+        var_z = self.extractor.extract(pc_in, [neighborhood], None, None, None)[0]
         np.testing.assert_allclose(var_z, 1.8408359999999995)
 
     def test_height_stats_without_neighbors(self):
         pc_in = load("testdata/AHN2.las")
         neighborhood = []
-        var_z = self.extractor.extract(pc_in, neighborhood, pc_in, None, None)
+        var_z = self.extractor.extract(pc_in, [neighborhood], pc_in, None, None)[0]
         assert np.isnan(var_z)
 
     def test_default_provides_correct(self):
@@ -38,7 +38,7 @@ class TestVarianceNormZFeatureExtractor(unittest.TestCase):
         point_cloud = create_point_cloud(x, y, z, normalized_z=normalized_z)
         neighborhood = [[0, 1, 2]]
 
-        variance = self.extractor.extract(point_cloud, neighborhood, None, None, None)
+        variance = self.extractor.extract(point_cloud, [neighborhood], None, None, None)[0]
 
         self.assertAlmostEqual(variance, 2 / 3)