+++ CHANGELOG.md
@@ -10,6 +10,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - the most relevant functions can be now imported directly from laserchicken
 
 ## Changed:
+- compute neighborhoods returns generator with neighborhoods instead of nested neighborhoods like it did before (breaking change!)
 - Some of the existing modules have been renamed/restructured (breaking changes!):
     - `normalization` --> `normalize`
     - `feature_extraction` created (functions moved from `feature_extractor/__init__.py`)